+++ app/fee.go
@@ -5,7 +5,8 @@ import (
 
 	"github.com/maxonrow/maxonrow-go/x/bank"
 	"github.com/maxonrow/maxonrow-go/x/fee"
-	token "github.com/maxonrow/maxonrow-go/x/token/fungible"
+	ft "github.com/maxonrow/maxonrow-go/x/token/fungible"
+	nft "github.com/maxonrow/maxonrow-go/x/token/nonfungible"
 
 	sdkTypes "github.com/cosmos/cosmos-sdk/types"
 	sdkAuth "github.com/cosmos/cosmos-sdk/x/auth"
@@ -38,15 +39,22 @@ func (app *mxwApp) CalculateFee(ctx sdkTypes.Context, tx sdkTypes.Tx) (sdkTypes.
 		var feeSetting *fee.FeeSetting
 
 		var isCustomAction = func(msg string) bool {
-			return msg == token.MsgTypeTransferFungibleToken ||
-				msg == token.MsgTypeMintFungibleToken ||
-				msg == token.MsgTypeBurnFungibleToken ||
-				msg == token.MsgTypeTransferFungibleTokenOwnership ||
-				msg == token.MsgTypeAcceptFungibleTokenOwnership
+			return msg == ft.MsgTypeTransferFungibleToken ||
+				msg == ft.MsgTypeMintFungibleToken ||
+				msg == ft.MsgTypeBurnFungibleToken ||
+				msg == ft.MsgTypeTransferFungibleTokenOwnership ||
+				msg == ft.MsgTypeAcceptFungibleTokenOwnership ||
+				// nft's
+				msg == nft.MsgTypeTransferNonFungibleItem ||
+				msg == nft.MsgTypeMintNonFungibleItem ||
+				msg == nft.MsgTypeBurnNonFungibleItem ||
+				msg == nft.MsgTypeTransferNonFungibleTokenOwnership ||
+				msg == nft.MsgTypeAcceptNonFungibleTokenOwnership ||
+				msg == nft.MsgTypeEndorsement
 		}
 		r := msg.Route()
 		t := msg.Type()
-		if r == token.MsgRoute &&
+		if r == ft.MsgRoute &&
 			isCustomAction(t) {
 
 			tokenFeeSetting, tokenAmt, feeSettingErr := app.getTokenFeeSetting(msg, ctx)
@@ -54,7 +62,22 @@ func (app *mxwApp) CalculateFee(ctx sdkTypes.Context, tx sdkTypes.Tx) (sdkTypes.
 				return nil, feeSettingErr
 			}
 
-			multiplier, multiplierErr = app.feeKeeper.GetTokenFeeMultiplier(ctx)
+			multiplier, multiplierErr = app.feeKeeper.GetFungibleTokenFeeMultiplier(ctx)
+			if multiplierErr != nil {
+				return nil, sdkTypes.ErrInternal("Get fee multiplier failed.")
+			}
+
+			amt = tokenAmt
+			feeSetting = tokenFeeSetting
+
+		} else if r == nft.MsgRoute &&
+			isCustomAction(t) {
+			tokenFeeSetting, tokenAmt, feeSettingErr := app.getTokenFeeSetting(msg, ctx)
+			if feeSettingErr != nil {
+				return nil, feeSettingErr
+			}
+
+			multiplier, multiplierErr = app.feeKeeper.GetNonFungibleTokenFeeMultiplier(ctx)
 			if multiplierErr != nil {
 				return nil, sdkTypes.ErrInternal("Get fee multiplier failed.")
 			}
@@ -133,9 +156,8 @@ func (app *mxwApp) getTokenFeeSetting(msg sdkTypes.Msg, ctx sdkTypes.Context) (*
 	var feeSettingErr sdkTypes.Error
 
 	switch msgType := msg.(type) {
-	case token.MsgTransferFungibleToken:
-
-		feeSetting, feeSettingErr = app.feeKeeper.GetTokenFeeSetting(ctx, msgType.Symbol, fee.TransferFungibleToken)
+	case ft.MsgTransferFungibleToken:
+		feeSetting, feeSettingErr = app.feeKeeper.GetFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.TransferToken)
 		if feeSettingErr != nil {
 			return nil, nil, feeSettingErr
 		}
@@ -149,8 +171,8 @@ func (app *mxwApp) getTokenFeeSetting(msg sdkTypes.Msg, ctx sdkTypes.Context) (*
 
 		amt = transferAmtCoins
 
-	case token.MsgMintFungibleToken:
-		feeSetting, feeSettingErr = app.feeKeeper.GetTokenFeeSetting(ctx, msgType.Symbol, fee.MintFungibleToken)
+	case ft.MsgMintFungibleToken:
+		feeSetting, feeSettingErr = app.feeKeeper.GetFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.MintToken)
 		if feeSettingErr != nil {
 			return nil, nil, feeSettingErr
 		}
@@ -163,8 +185,8 @@ func (app *mxwApp) getTokenFeeSetting(msg sdkTypes.Msg, ctx sdkTypes.Context) (*
 
 		amt = mintAmtCoins
 
-	case token.MsgBurnFungibleToken:
-		feeSetting, feeSettingErr = app.feeKeeper.GetTokenFeeSetting(ctx, msgType.Symbol, fee.BurnFungibleToken)
+	case ft.MsgBurnFungibleToken:
+		feeSetting, feeSettingErr = app.feeKeeper.GetFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.BurnToken)
 		if feeSettingErr != nil {
 			return nil, nil, feeSettingErr
 		}
@@ -177,14 +199,51 @@ func (app *mxwApp) getTokenFeeSetting(msg sdkTypes.Msg, ctx sdkTypes.Context) (*
 
 		amt = burnAmtCoins
 
-	case token.MsgTransferFungibleTokenOwnership:
-		feeSetting, feeSettingErr = app.feeKeeper.GetTokenFeeSetting(ctx, msgType.Symbol, fee.TransferTokenOwnership)
+	case ft.MsgTransferFungibleTokenOwnership:
+		feeSetting, feeSettingErr = app.feeKeeper.GetFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.TransferTokenOwnership)
 		if feeSettingErr != nil {
 			return nil, nil, feeSettingErr
 		}
 
-	case token.MsgAcceptFungibleTokenOwnership:
-		feeSetting, feeSettingErr = app.feeKeeper.GetTokenFeeSetting(ctx, msgType.Symbol, fee.AcceptTokenOwnership)
+	case ft.MsgAcceptFungibleTokenOwnership:
+		feeSetting, feeSettingErr = app.feeKeeper.GetFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.AcceptTokenOwnership)
+		if feeSettingErr != nil {
+			return nil, nil, feeSettingErr
+		}
+
+	// nft's
+	case nft.MsgTransferNonFungibleItem:
+		feeSetting, feeSettingErr = app.feeKeeper.GetNonFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.TransferToken)
+		if feeSettingErr != nil {
+			return nil, nil, feeSettingErr
+		}
+
+	case nft.MsgMintNonFungibleItem:
+		feeSetting, feeSettingErr = app.feeKeeper.GetNonFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.MintToken)
+		if feeSettingErr != nil {
+			return nil, nil, feeSettingErr
+		}
+
+	case nft.MsgBurnNonFungibleItem:
+		feeSetting, feeSettingErr = app.feeKeeper.GetNonFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.BurnToken)
+		if feeSettingErr != nil {
+			return nil, nil, feeSettingErr
+		}
+
+	case nft.MsgTransferNonFungibleTokenOwnership:
+		feeSetting, feeSettingErr = app.feeKeeper.GetNonFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.TransferTokenOwnership)
+		if feeSettingErr != nil {
+			return nil, nil, feeSettingErr
+		}
+
+	case nft.MsgAcceptNonFungibleTokenOwnership:
+		feeSetting, feeSettingErr = app.feeKeeper.GetNonFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.AcceptTokenOwnership)
+		if feeSettingErr != nil {
+			return nil, nil, feeSettingErr
+		}
+
+	case nft.MsgEndorsement:
+		feeSetting, feeSettingErr = app.feeKeeper.GetNonFungibleTokenFeeSetting(ctx, msgType.Symbol, fee.Endorse)
 		if feeSettingErr != nil {
 			return nil, nil, feeSettingErr
 		}