+++ x/token/fungible/keeper.go
@@ -336,7 +336,7 @@ func (k *Keeper) approveFungibleToken(ctx sdkTypes.Context, symbol string, token
 		if !k.feeKeeper.FeeSettingExists(ctx, tokenFee.FeeName) {
 			return types.ErrFeeSettingNotExists(tokenFee.FeeName).Result()
 		}
-		err := k.feeKeeper.AssignFeeToTokenAction(ctx, tokenFee.FeeName, token.Symbol, tokenFee.Action)
+		err := k.feeKeeper.AssignFeeToFungibleTokenAction(ctx, tokenFee.FeeName, token.Symbol, tokenFee.Action)
 		if err != nil {
 			return err.Result()
 		}